% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-JTW_1991_SYD.R
\docType{data}
\name{JTW_1991_SYD}
\alias{JTW_1991_SYD}
\title{Journey to work data, 1991, Sydney}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 3208 rows and 3 columns.}
\source{
\url{https://opendata.transport.nsw.gov.au/dataset/journey-work-jtw-1991}
}
\usage{
JTW_1991_SYD
}
\description{
Counts of workers by home and work SLA from the 1991 census. Courtesy of Transport for NSW.
}
\keyword{datasets}
