% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-JTW_2001_TZ__NSW_GMA.R
\docType{data}
\name{JTW_2001_TZ__NSW_GMA}
\alias{JTW_2001_TZ__NSW_GMA}
\title{2001 NSW journey to work data by travel zone}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 196334 rows and 11 columns.}
\source{
\url{https://opendata.transport.nsw.gov.au/dataset/travel-zones-2006}
}
\usage{
JTW_2001_TZ__NSW_GMA
}
\description{
2001 journey to work data, by travel zone.
}
\keyword{datasets}
