% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-JTW_2006_SLA.R
\docType{data}
\name{JTW_2006_SLA}
\alias{JTW_2006_SLA}
\title{2006 journey to work data, by SLA}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2085055 rows and 3 columns.}
\usage{
JTW_2006_SLA
}
\description{
Origin-destination counts from the journey to work question in the Australian census.

"Q41: For the main job held last week, what was the person's work place address?
\itemize{
\item{For persons who usually worked from home, write the home address}
\item{For persons with no fixed place of work:}
\itemize{
\item{If the person usually travels to a depot to start work, write the depot address;}
\item{otherwise, write \code{NONE} in 'Suburb/Locality' box.}
}
}
}
\keyword{datasets}
